requirements.txt
yfinance
supabase
openai
python-dotenv
pandas
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
beautifulsoup4
requests